Fix broken password property and flag
In previous releases the password set via the config file or command
line was ignored. If the keyring was enabled it was used, if it was
not set then user was prompted for the password.

Now the password property and flag are respected. If the keyring
is enabled it is used, and the password property and flag are still
ignored. If the keyring is enabled, but not yet set then the user
is prompted even if a password is given via the password property
or flag. The keyring is preferred because it is more secure storage.