Add CHANGELOG.rst (#74)
The CHANGELOG is in reStructuredText and stylized as a release 
history so that it can be reused as is in Sphinx documentation. For
that reason it is only loosely based on the Keep a Changelog markdown
format.

The version links in the example from Keep a Changelog link to
Github diffs. The links in the CHANGELOG I authored link to the
corresponding release on PyPI which seems much more useful to end
users.