Add invalid selection error message (#71)
In previous releases an invalid selection caused a traceback to be
printed to the user. In the current release no error message is
displayed. With this commit an error message is issued to the user
on an invalid selection. This is sufficient since the user does not
need to reauthenticate on additional login attempts.