Use date for age based cache invalidation
The script weekold.py was used to invalidate a cache after a week.
The script worked but the cache contents were not updated since
cache contents are saved only on a cache miss. The only way to
update a Github Actions cache is to change the key.

This commit removes the weekold.py script, and inserts the week
number in the cache key name. This ensures that the cache can never
be more than a week old.